SAR image denoise program
how can i succesed in doing this kind of job? this is my rencent work:in
this program,i put up my research to public ,but in irder to keep
serect,some changes had been made ,it seems like too much time to run
it,the last but not least the idea matters.